# api.yaml
openapi: 3.0.0
info:
  title: "Szabadságnyilvántartó API"
  version: 1.0.0
paths:
...
  /szabadsagok/own:
    {
      get:
        {
          summary: "Retrieve the szabadsagok of the authenticated user",
          tags: [Szabadsagok],
          security: [{ ApiKeyAuth: [] }],
          responses:
            {
              "200":
                {
                  description: "Returns a status and a list of szabadsagok or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  szabadsagok:
                                    {
                                      type: array,
                                      items:
                                        {
                                          $ref: "#/components/schemas/Szabadsag",
                                        },
                                    },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  "/szabadsagok/{id}":
    {
      get:
        {
          summary: "Retrieve the szabadsagok of a user by ID",
          tags: [Szabadsagok],
          security: [{ ApiKeyAuth: [] }],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the user whose szabadsagok to retrieve",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns a status and a list of szabadsagok or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  szabadsag:
                                    {
                                      type: array,
                                      items:
                                        {
                                          $ref: "#/components/schemas/Szabadsag",
                                        },
                                    },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
...
components:
  securitySchemes:
    { ApiKeyAuth: { type: apiKey, in: header, name: submittingId } }
  schemas:
...
      Szabadsag:
        {
          type: object,
          properties:
            {
              type: { type: string, default: sz },
              dates: { type: array, items: { format: date } },
              userId: { type: string },
              managerId: { type: string },
            },
        },
...